feat: E2E 테스트 생성 후 패키지로 분리
