--- src/test/java/roomescape/MissionStep3Test.java
@@ -1,74 +1,0 @@
-package roomescape;
-
-import static org.hamcrest.Matchers.is;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import java.util.HashMap;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class MissionStep3Test {
-
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @Test
-    void 시간_관리_API() {
-        Map<String, String> params = new HashMap<>();
-        params.put("startAt", "20:00");
-        params.put("finishAt", "21:00");
-
-        Integer newId = RestAssured.given().log().all()
-            .contentType(ContentType.JSON)
-            .body(params)
-            .when().post("/times")
-            .then().log().all()
-            .statusCode(201)
-            .extract().path("id");
-
-        RestAssured.given().log().all()
-            .when().get("/times")
-            .then().log().all()
-            .statusCode(200)
-            .body("size()", is(4));
-
-        RestAssured.given().log().all()
-            .when().delete("/times/" + newId)
-            .then().log().all()
-            .statusCode(204);
-    }
-
-    @Test
-    void 예약과_시간_연결() {
-        Map<String, Object> reservation = new HashMap<>();
-        reservation.put("name", "브라운");
-        reservation.put("date", "2099-08-05");
-        reservation.put("timeId", 1);
-        reservation.put("themeId", 1);
-
-        RestAssured.given().log().all()
-            .contentType(ContentType.JSON)
-            .body(reservation)
-            .when().post("/reservations")
-            .then().log().all()
-            .statusCode(201);
-
-        RestAssured.given().log().all()
-            .when().get("/times/available?date=2099-08-05&themeId=1")
-            .then().log().all()
-            .statusCode(200)
-            .body("size()", is(2));
-    }
-}
