+++ src/test/java/roomescape/e2e/MissionStep2Test.java
@@ -1,14 +1,15 @@
-package roomescape;
+package roomescape.e2e;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import java.sql.Connection;
-import java.sql.SQLException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
