+++ src/test/java/roomescape/e2e/MissionStep4Test.java
@@ -1,14 +1,15 @@
-package roomescape;
+package roomescape.e2e;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import java.lang.reflect.Field;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.reservation.controller.ReservationController;
+
+import java.lang.reflect.Field;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
